CIT-739 refactor(tests): reorganize dictionary fixtures and update imports
