+++ tests/conftest.py
@@ -8,12 +8,14 @@ from typing import TYPE_CHECKING, Callable, Optional, Union
 import numpy as np
 import pytest
 from ingenialink import Servo
+from ingenialink.dictionary import Interface
 from ingenialink.exceptions import ILRegisterNotFoundError
 from summit_testing_framework import dynamic_loader
 from summit_testing_framework.profilers.stoppable_gaps import StoppableProfilerConfig
 from summit_testing_framework.pytest_helpers.marker_helper import (
     apply_firmware_version_markers_to_items,
 )
+from summit_testing_framework.setups.specifiers import DictionaryType, DictionaryVersion
 
 if TYPE_CHECKING:
     from ingeniamotion.axis import Axis
@@ -215,3 +217,23 @@ def refresh_registers_for_test_rollback(servo: Servo, register_uids: list[str]):
             logger.warning(
                 f"Register {register_uid} not found during refresh after test execution."
             )
+
+
+@pytest.fixture(scope="session")
+def sample_safe_ph1_xdfv3_dictionary(
+    product_dictionary: Callable[[str, DictionaryVersion, Optional[Interface]], Path],
+) -> Path:
+    return product_dictionary(
+        "DEN-S-NET-E",
+        DictionaryVersion("2.9.1", DictionaryType.XDF_V3),
+    )
+
+
+@pytest.fixture(scope="session")
+def sample_safe_ph2_xdfv3_dictionary(
+    product_dictionary: Callable[[str, DictionaryVersion, Optional[Interface]], Path],
+) -> Path:
+    return product_dictionary(
+        "EVS-S-NET-E",
+        DictionaryVersion("2.9.1", DictionaryType.XDF_V3),
+    )
