+++ tests/fsoe/test_process_image.py
@@ -10,12 +10,8 @@ from ingenialink.pdo import RPDOMap, TPDOMap
 from ingeniamotion.enums import FSoEState
 from ingeniamotion.fsoe import FSOE_MASTER_INSTALLED, FSoEError
 from ingeniamotion.motion_controller import MotionController
-from tests.fsoe.conftest import (
-    MockNetwork,
-    MockServo,
-    sample_safe_ph1_xdfv3_dictionary,
-    sample_safe_ph2_xdfv3_dictionary,
-)
+from tests.conftest import sample_safe_ph1_xdfv3_dictionary, sample_safe_ph2_xdfv3_dictionary
+from tests.fsoe.conftest import MockNetwork, MockServo
 
 try:
     import pysoem
